third_party/blink/web_tests/platform/generic/fast/js/function-argument-evaluation-expected.txt

This page tests function calls whose argument expressions overwrite the callee.

If the test passes, you'll see PASS messages below.

PASS: test1(callback, 1) should be 1 and is.
PASS: test2(callback, 1) should be 1 and is.
